Remove -logtodebugger, minor code cleanup.
"-logtodebugger" is a strange, obscure, WIN32-only (mostly MSVC) thing.
Let's clean up the options a bit get rid of it.

test_bitcoin was using fLogToDebugger as a way to prevent logging to debug.log.
fLogToDebugger now replaced with a boolean (not exposed as option) fLogToDebugLog that defaults to true.

https://github.com/bitcoin/bitcoin/commit/9e9056cd1a05e5868025553d2a8cb3e47cf45045